Improved buttons for accounts prompt

+++ src/main/java/me/gosdev/chatpointsttv/Commands/AccountsCommand.java
@@ -1,5 +1,6 @@
 package me.gosdev.chatpointsttv.Commands;
 
+import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 
 import me.gosdev.chatpointsttv.ChatPointsTTV;
@@ -19,27 +20,32 @@ public class AccountsCommand {
             return;
         }
         TextComponent msg = new TextComponent("---------- " + ChatColor.LIGHT_PURPLE + ChatColor.BOLD  + "Connected Accounts" + ChatColor.RESET + " ----------\n\n");
+        if (p.equals(Bukkit.getConsoleSender())) {
 
-        for (Channel channel : ChatPointsTTV.getPlugin().getTwitch().getListenedChannels().values()) {
-            BaseComponent deleteButton = new ComponentBuilder(ChatColor.RED + "  [x]").create()[0];
-            deleteButton.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to unlink this account").create()));
-            deleteButton.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/twitch unlink " + channel.getChannelUsername()));
-            msg.addExtra(deleteButton);
-            msg.addExtra(new TextComponent(" " + channel.getChannelUsername() + "\n"));
+        } else {
+            for (Channel channel : ChatPointsTTV.getPlugin().getTwitch().getListenedChannels().values()) {
+                BaseComponent deleteButton = new ComponentBuilder(ChatColor.RED + "  [❌]").create()[0];
+                deleteButton.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to unlink this account").create()));
+                deleteButton.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/twitch unlink " + channel.getChannelUsername()));
+                msg.addExtra(deleteButton);
+                msg.addExtra(new TextComponent("  " + channel.getChannelUsername() + "\n"));
+            }
+    
+            TextComponent addBtn = new TextComponent(ChatColor.GREEN + "" + ChatColor.BOLD + "\n[+]" + ChatColor.RESET + ChatColor.GREEN + " Add account");
+            addBtn.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to link another account").create()));
+            addBtn.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/twitch link"));
+    
+            TextComponent stopBtn = new TextComponent(ChatColor.RED + "" + ChatColor.BOLD + "[❌]" + ChatColor.RESET + ChatColor.RED + " Remove all accounts");
+            stopBtn.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to unlink all accounts").create()));
+            stopBtn.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/twitch unlink"));
+    
+            msg.addExtra(addBtn);
+            msg.addExtra(ChatColor.GRAY + "  -  ");
+            msg.addExtra(stopBtn);
+    
+    
         }
-
-        TextComponent addBtn = new TextComponent(ChatColor.GREEN + "" + ChatColor.BOLD + "\n[+]" + ChatColor.RESET + ChatColor.GREEN + " Add account");
-        addBtn.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to link another account").create()));
-        addBtn.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/twitch link"));
-
-        TextComponent stopBtn = new TextComponent(ChatColor.RED + "" + ChatColor.BOLD + "[x]" + ChatColor.RESET + ChatColor.RED + " Remove all accounts");
-        stopBtn.setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new ComponentBuilder("Click to unlink all accounts").create()));
-        stopBtn.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/twitch unlink"));
-
-        msg.addExtra(addBtn);
-        msg.addExtra(ChatColor.GRAY + " - ");
-        msg.addExtra(stopBtn);
-
         p.spigot().sendMessage(msg);
+
     }
 }
